Fix wrappers for custom elements
We now associate the prototype object with the wrapper constructor instead of
the native constructor. This is better because constructor is not as reliable as
using the prototype.

Generate the wrapper constructor lazily. This allows the wrapper constructor to
be generated on the fly based on the prototype chain which is required for
custom elements.

Use __proto__ as a hint for IE which does not have __proto__.