Fix issue with event parents
Turns out we cannot store the event parents on the insertion point, since the
insertion point might have different event parent depending on the where the
event originated from. Instead we store the event parents on the distributed
node as an array and shift off of that when we get the event path during
dispatch.